requirements.txt
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
Django==5.2
django-extensions==3.2.3
django-filter==25.1
django-silk==5.3.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gprof2dot==2024.6.6
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
pillow==11.1.0
pycodestyle==2.13.0
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.2.3
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
